==> part1/ch2_end_to_end_machine_learning_project/5_data_preprocessing.ipynb <==
# 2.5 머신러닝 알고리즘을 위한 데이터 준비

import matplotlib.pyplot as plt
import pandas as pd
import sklearn
import numpy as np

import os
HOUSING_PATH = os.path.join("datasets", "housing")
def load_housing_data(housing_path=HOUSING_PATH):
    csv_path = os.path.join(housing_path, "housing.csv")
    return pd.read_csv(csv_path)

housing=load_housing_data()

housing["income_cat"] = np.ceil(housing["median_income"] / 1.5)
housing["income_cat"].where(housing["income_cat"] < 5, 5.0, inplace=True)

from sklearn.model_selection import StratifiedShuffleSplit
split=StratifiedShuffleSplit(n_splits=1,test_size=0.2,random_state=42)

for train_index, test_index in split.split(housing, housing["income_cat"]):
    strat_train_set = housing.loc[train_index]
    strat_test_set = housing.loc[test_index]

### 2.5.0 예측 변수와 레이블 분리

# drop() 함수는 해당 에트리뷰트를 제외하고 데이터를 복사한다
housing=strat_train_set.drop("median_house_value",axis=1)
housing.head()

# 훈련 데이터셋에서 "median_house_value" 에트리뷰트만 분리하여 복사
housing_labels=strat_train_set["median_house_value"].copy()
housing_labels=pd.DataFrame(housing_labels)
housing_labels.to_csv("datasets/housing/housing_labels.csv", mode='w')
housing_labels.head()

### 2.5.1 데이터 정제
- 누락된 값을 처리할 수 있는 함수를 만든다
- total bedrooms 변수에는 누락된 값이 존재한다

housing.info()

#### 누락된 값이 있는 열 제거

housing.dropna(subset=["total_bedrooms"]) # 이거 왜 안돼지?????
housing.info()

#### 전체 특성(에트리뷰트) 삭제

housing=strat_train_set.drop("median_house_value",axis=1)# 데이터 복원
housing.drop("total_bedrooms",axis=1) # 이건 또 왜 안돼
housing.info()

#### 특정 값으로 채우기

housing=strat_train_set.drop("median_house_value",axis=1)# 데이터 복원
median=housing["total_bedrooms"].median() # 해당 에트리뷰트 중간값
housing["total_bedrooms"].fillna(median,inplace=True) # 중간값으로 누락된 값 채움
housing.info()

#### 사이킷런의 Imputer 사용

housing=strat_train_set.drop("median_house_value",axis=1)# 데이터 복원
housing_num=housing.drop("ocean_proximity",axis=1) # 텍스트 타입 데이터 제거

from sklearn.impute import SimpleImputer
imputer=SimpleImputer(strategy="median") # 누락된 값을 중간값으로 채우는 imputer 생성
imputer.fit(housing_num) # imputer를 훈련 데이터에 적용
X=imputer.transform(housing_num) # imputer를 적용하여 변환, 넘파이 배열
housing_tr=pd.DataFrame(X,columns=housing_num.columns,index=list(housing.index.values)) # 다시 pandas의 dataframe 타입으로 전환

housing_tr.info()

print(imputer.statistics_) # imputer는 각 특성의 중간값을 계산하여 statistics_ 변수에 저장한다
print(housing_num.median().values) # 숫자타입 에트리뷰트 중간값 배열

### 2.5.2 텍스트와 범주형 특성 다루기

housing_cat=housing[["ocean_proximity"]]  # 괄호 두개는 열벡터..? dataframe?
housing_cat.head()

#### 방법 1 - OrdinalEncoder + OneHotEncoder

# from future_encoders import OrdinalEncoder
from sklearn.preprocessing import OrdinalEncoder
ordinal_encoder = OrdinalEncoder()
housing_cat_encoded = ordinal_encoder.fit_transform(housing_cat)
housing_cat_encoded[:10]

# sklearn 의 One Hot Incoder 클래스 사용
from sklearn.preprocessing import OneHotEncoder
encoder=OneHotEncoder(categories='auto')
housing_cat_1hot=encoder.fit_transform(housing_cat_encoded.reshape(-1,1))
housing_cat_1hot # scipy 희소행렬
housing_cat_1hot.toarray() # numpy array 타입으로 변환

# sparse=false로 지정하면 OrdinalEncoder 필요 없음
cat_encoder = OneHotEncoder(categories='auto', sparse=False)
housing_cat_1hot = cat_encoder.fit_transform(housing_cat)
housing_cat_1hot

#### 방법 2 - CategoricalEncoder

from sklearn.preprocessing import CategoricalEncoder 

cat_encoder = CategoricalEncoder()
housing_cat_reshaped = housing_cat.values.reshape(-1, 1) # 열 벡터로 바꿈
housing_cat_1hot = cat_encoder.fit_transform(housing_cat_reshaped)
housing_cat_1hot

sklearn.__version__ # 왜 임포트가 안되는가... >> 0.20에서만 된다

# [PR #9151](https://github.com/scikit-learn/scikit-learn/pull/9151)에서 가져온 CategoricalEncoder 클래스의 정의.
# 이 클래스는 사이킷런 0.20에 포함될 예정입니다.

from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.utils import check_array
from sklearn.preprocessing import LabelEncoder
from scipy import sparse

class CategoricalEncoder(BaseEstimator, TransformerMixin):
    """Encode categorical features as a numeric array.
    The input to this transformer should be a matrix of integers or strings,
    denoting the values taken on by categorical (discrete) features.
    The features can be encoded using a one-hot aka one-of-K scheme
    (``encoding='onehot'``, the default) or converted to ordinal integers
    (``encoding='ordinal'``).
    This encoding is needed for feeding categorical data to many scikit-learn
    estimators, notably linear models and SVMs with the standard kernels.
    Read more in the :ref:`User Guide <preprocessing_categorical_features>`.
    Parameters
    ----------
    encoding : str, 'onehot', 'onehot-dense' or 'ordinal'
        The type of encoding to use (default is 'onehot'):
        - 'onehot': encode the features using a one-hot aka one-of-K scheme
          (or also called 'dummy' encoding). This creates a binary column for
          each category and returns a sparse matrix.
        - 'onehot-dense': the same as 'onehot' but returns a dense array
          instead of a sparse matrix.
        - 'ordinal': encode the features as ordinal integers. This results in
          a single column of integers (0 to n_categories - 1) per feature.
    categories : 'auto' or a list of lists/arrays of values.
        Categories (unique values) per feature:
        - 'auto' : Determine categories automatically from the training data.
        - list : ``categories[i]`` holds the categories expected in the ith
          column. The passed categories are sorted before encoding the data
          (used categories can be found in the ``categories_`` attribute).
    dtype : number type, default np.float64
        Desired dtype of output.
    handle_unknown : 'error' (default) or 'ignore'
        Whether to raise an error or ignore if a unknown categorical feature is
        present during transform (default is to raise). When this is parameter
        is set to 'ignore' and an unknown category is encountered during
        transform, the resulting one-hot encoded columns for this feature
        will be all zeros.
        Ignoring unknown categories is not supported for
        ``encoding='ordinal'``.
    Attributes
    ----------
    categories_ : list of arrays
        The categories of each feature determined during fitting. When
        categories were specified manually, this holds the sorted categories
        (in order corresponding with output of `transform`).
    Examples
    --------
    Given a dataset with three features and two samples, we let the encoder
    find the maximum value per feature and transform the data to a binary
    one-hot encoding.
    >>> from sklearn.preprocessing import CategoricalEncoder
    >>> enc = CategoricalEncoder(handle_unknown='ignore')
    >>> enc.fit([[0, 0, 3], [1, 1, 0], [0, 2, 1], [1, 0, 2]])
    ... # doctest: +ELLIPSIS
    CategoricalEncoder(categories='auto', dtype=<... 'numpy.float64'>,
              encoding='onehot', handle_unknown='ignore')
    >>> enc.transform([[0, 1, 1], [1, 0, 4]]).toarray()
    array([[ 1.,  0.,  0.,  1.,  0.,  0.,  1.,  0.,  0.],
           [ 0.,  1.,  1.,  0.,  0.,  0.,  0.,  0.,  0.]])
    See also
    --------
    sklearn.preprocessing.OneHotEncoder : performs a one-hot encoding of
      integer ordinal features. The ``OneHotEncoder assumes`` that input
      features take on values in the range ``[0, max(feature)]`` instead of
      using the unique values.
    sklearn.feature_extraction.DictVectorizer : performs a one-hot encoding of
      dictionary items (also handles string-valued features).
    sklearn.feature_extraction.FeatureHasher : performs an approximate one-hot
      encoding of dictionary items or strings.
    """

    def __init__(self, encoding='onehot', categories='auto', dtype=np.float64,
                 handle_unknown='error'):
        self.encoding = encoding
        self.categories = categories
        self.dtype = dtype
        self.handle_unknown = handle_unknown

    def fit(self, X, y=None):
        """Fit the CategoricalEncoder to X.
        Parameters
        ----------
        X : array-like, shape [n_samples, n_feature]
            The data to determine the categories of each feature.
        Returns
        -------
        self
        """

        if self.encoding not in ['onehot', 'onehot-dense', 'ordinal']:
            template = ("encoding should be either 'onehot', 'onehot-dense' "
                        "or 'ordinal', got %s")
            raise ValueError(template % self.handle_unknown)

        if self.handle_unknown not in ['error', 'ignore']:
            template = ("handle_unknown should be either 'error' or "
                        "'ignore', got %s")
            raise ValueError(template % self.handle_unknown)

        if self.encoding == 'ordinal' and self.handle_unknown == 'ignore':
            raise ValueError("handle_unknown='ignore' is not supported for"
                             " encoding='ordinal'")

        X = check_array(X, dtype=np.object, accept_sparse='csc', copy=True)
        n_samples, n_features = X.shape

        self._label_encoders_ = [LabelEncoder() for _ in range(n_features)]

        for i in range(n_features):
            le = self._label_encoders_[i]
            Xi = X[:, i]
            if self.categories == 'auto':
                le.fit(Xi)
            else:
                valid_mask = np.in1d(Xi, self.categories[i])
                if not np.all(valid_mask):
                    if self.handle_unknown == 'error':
                        diff = np.unique(Xi[~valid_mask])
                        msg = ("Found unknown categories {0} in column {1}"
                               " during fit".format(diff, i))
                        raise ValueError(msg)
                le.classes_ = np.array(np.sort(self.categories[i]))

        self.categories_ = [le.classes_ for le in self._label_encoders_]

        return self

    def transform(self, X):
        """Transform X using one-hot encoding.
        Parameters
        ----------
        X : array-like, shape [n_samples, n_features]
            The data to encode.
        Returns
        -------
        X_out : sparse matrix or a 2-d array
            Transformed input.
        """
        X = check_array(X, accept_sparse='csc', dtype=np.object, copy=True)
        n_samples, n_features = X.shape
        X_int = np.zeros_like(X, dtype=np.int)
        X_mask = np.ones_like(X, dtype=np.bool)

        for i in range(n_features):
            valid_mask = np.in1d(X[:, i], self.categories_[i])

            if not np.all(valid_mask):
                if self.handle_unknown == 'error':
                    diff = np.unique(X[~valid_mask, i])
                    msg = ("Found unknown categories {0} in column {1}"
                           " during transform".format(diff, i))
                    raise ValueError(msg)
                else:
                    # Set the problematic rows to an acceptable value and
                    # continue `The rows are marked `X_mask` and will be
                    # removed later.
                    X_mask[:, i] = valid_mask
                    X[:, i][~valid_mask] = self.categories_[i][0]
            X_int[:, i] = self._label_encoders_[i].transform(X[:, i])

        if self.encoding == 'ordinal':
            return X_int.astype(self.dtype, copy=False)

        mask = X_mask.ravel()
        n_values = [cats.shape[0] for cats in self.categories_]
        n_values = np.array([0] + n_values)
        indices = np.cumsum(n_values)

        column_indices = (X_int + indices[:-1]).ravel()[mask]
        row_indices = np.repeat(np.arange(n_samples, dtype=np.int32),
                                n_features)[mask]
        data = np.ones(n_samples * n_features)[mask]

        out = sparse.csc_matrix((data, (row_indices, column_indices)),
                                shape=(n_samples, indices[-1]),
                                dtype=self.dtype).tocsr()
        if self.encoding == 'onehot-dense':
            return out.toarray()
        else:
            return out

### 2.5.3 나만의 변환기 만들기

from sklearn.base import BaseEstimator, TransformerMixin

# 컬럼 인덱스
rooms_ix, bedrooms_ix, population_ix, household_ix = 3, 4, 5, 6

class CombinedAttributesAdder(BaseEstimator, TransformerMixin):
    def __init__(self, add_bedrooms_per_room = True): # no *args or **kargs
        self.add_bedrooms_per_room = add_bedrooms_per_room
    def fit(self, X, y=None):
        return self  # nothing else to do
    def transform(self, X, y=None):
        rooms_per_household = X[:, rooms_ix] / X[:, household_ix] # 가구당 방 갯수
        population_per_household = X[:, population_ix] / X[:, household_ix] # 가구당 사람 수
        if self.add_bedrooms_per_room:
            bedrooms_per_room = X[:, bedrooms_ix] / X[:, rooms_ix] #방 갯수대비 침실 갯수
            return np.c_[X, rooms_per_household, population_per_household,
                         bedrooms_per_room]
        else:
            return np.c_[X, rooms_per_household, population_per_household] # 1차원 배열을 새로로 붙이기

attr_adder = CombinedAttributesAdder(add_bedrooms_per_room=False)
housing_extra_attribs = attr_adder.transform(housing.values)

housing_extra_attribs = pd.DataFrame( # 다시 dataframe 배열로 전환
    housing_extra_attribs, 
    columns=list(housing.columns)+["rooms_per_household", "population_per_household"])
housing_extra_attribs.head()

### 2.5.4 특성 스케일링 (feature scaling)
- 머신러닝 알고리즘은 입력 숫자 특성들의 스케일(범위)가 다르면 잘 작동하지 않을 수 있다
- normalization (min-max scaling) : 0~1사이 값으로 스케일 조정, MinMaxScaler 사용
- standadization : 평균 0, 표준편차 1이 되도록 스케일링, StandardScaler 사용

### 2.5.5 변환 파이프라인
- Pipeline 클래스 : 연속된 변환을 순서대로 처리할 수 있음
- 연속된 단계를 나타내는 이름/추정기 쌍의 목록을 입력으로 받는다
- 마지막 단계를 제외한 나머지는 모두 변환기여야 한다. 즉 fit과 transform 또는 fit_transform 메소드를 가지고 있어야 한다.
- 마지막 단계는 변환기와 추정기를 모두 사용할 수 있다.

from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler

num_pipeline = Pipeline([
        ('imputer', SimpleImputer(strategy="median")), # 변환기 : 누락된 값 채워줌
        ('attribs_adder', CombinedAttributesAdder()), # 변환기 : 다른 특성 추가
        ('std_scaler', StandardScaler()), # 변환기 : 특성 스케일링
    ])

housing_num_tr = num_pipeline.fit_transform(housing_num)
housing_num_tr = pd.DataFrame(housing_num_tr,columns=list(housing_extra_attribs.columns))
housing_num_tr.head()

from sklearn.base import BaseEstimator, TransformerMixin

# 사이킷런이 DataFrame을 바로 사용하지 못하므로 수치형이나 범주형 컬럼을 선택하는 클래스를 만듭니다.

class DataFrameSelector(BaseEstimator, TransformerMixin):
    def __init__(self, attribute_names):
        self.attribute_names = attribute_names
    def fit(self, X, y=None):
        return self
    def transform(self, X):
        return X[self.attribute_names].values

num_attribs = list(housing_num)
cat_attribs = ["ocean_proximity"]

num_pipeline = Pipeline([
        ('selector', DataFrameSelector(num_attribs)),
        ('imputer', SimpleImputer(strategy="median")),
        ('attribs_adder', CombinedAttributesAdder()),
        ('std_scaler', StandardScaler()),
    ])

cat_pipeline = Pipeline([
        ('selector', DataFrameSelector(cat_attribs)),
        ('cat_encoder', CategoricalEncoder(encoding="onehot-dense")),
    ])

# 두 파이프라인을 하나의 파이프라인으로 합치기

from sklearn.pipeline import FeatureUnion

full_pipeline=FeatureUnion(transformer_list=[
    ("num_pipeline",num_pipeline),
    ("cat_pipeline",cat_pipeline),
])

housing_prepared = full_pipeline.fit_transform(housing)
housing_prepared

from sklearn.externals import joblib
joblib.dump(full_pipeline, "full_pipeline.pkl")

# from future_encoders import ColumnTransformer
from sklearn.compose import ColumnTransformer

num_attribs = list(housing_num)
cat_attribs = ["ocean_proximity"]

full_pipeline = ColumnTransformer([
        ("num", num_pipeline, num_attribs), # num_pipeline에 num_attribs 적용
        ("cat", OneHotEncoder(categories='auto'), cat_attribs), # OneHotEncoder에 cat_attribs 적용
    ])

housing_prepared = full_pipeline.fit_transform(housing)
np.save("datasets/housing/housing_prepared.npy", housing_prepared)


# 요약
- 누락된 값을 처리할 때에는 imputer 사용
- 데이터를 객체에 적용할 때에는 fit, 데이터를 변환할 때에는 trasnform 메소드 사용, 또는 fit_transform 사용
- transform 이후에 넘파이 배열이 반환되면 pd.Dataframe을 이용하여 타입 변환 가능
- 텍스트형 데이터를 다룰 때에는 OrdinalIncoder, OneHotIncoder 사용
- 연속된 변환을 처리할 때에는 Pipeline, 파이프라인을 함칠 때에는 FeatureUnion
