==> part1/ch2_end_to_end_machine_learning_project/6_model_selection_and_training.ipynb <==
# 2.6 모델 선택과 훈련

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np

#load_data
housing_prepared=np.load("datasets/housing/housing_prepared.npy")
housing_labels=pd.read_csv("datasets/housing/housing_labels.csv",usecols=[1],squeeze=True)
print(housing_labels.shape)
print(housing_prepared.shape)

### 2.6.1 훈련 세트에서 훈련하고 평가하기

from sklearn.linear_model import LinearRegression 
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor

lin_reg=LinearRegression()
lin_reg.fit(housing_prepared,housing_labels)
lin_reg_predictions=lin_reg.predict(housing_prepared)

tree_reg=DecisionTreeRegressor()
tree_reg.fit(housing_prepared,housing_labels)
tree_reg_predictions=tree_reg.predict(housing_prepared)

forest_reg=RandomForestRegressor()
forest_reg.fit(housing_prepared,housing_labels)
forest_reg_predictions=forest_reg.predict(housing_prepared)

from sklearn.metrics import mean_squared_error
def rmse(labels,predictions):
    mse=mean_squared_error(labels,predictions)
    rmse=np.sqrt(mse)
    return rmse

lin_rmse=rmse(housing_labels,lin_reg_predictions)
tree_rmse=rmse(housing_labels,tree_reg_predictions)
forest_rmse=rmse(housing_labels,forest_reg_predictions)

print(lin_rmse) # underfitting
print(tree_rmse) # overfitting
print(forest_rmse)

### 2. K-fold cross validation (교차검증)
- 훈련 세트를 폴드(fold)라고 불리는 n개의 서브셋으로 무작위 분할한다
- 매번 다른 폴드를 선택해 모델 평가에 사용하고 나머지 n-1개의 폴드는 훈련에 사용한다
- 교차검증은 비용함수가 아닌 효용함수를 사용한다

from sklearn.model_selection import cross_val_score
# 모델 객체, X, y, 점수 측정 방법, k 값(?)
scores=cross_val_score(tree_reg,housing_prepared,housing_labels,scoring="neg_mean_squared_error",cv=10)

print("scores :",scores) #???? 결과 왜 저럼??????????
print("mean :",scores.mean()) 
print("standard deviation :",scores.std())

scores_lin=cross_val_score(lin_reg,housing_prepared,housing_labels,scoring="neg_mean_squared_error",cv=10)
print("scores :",scores_lin) #???? 결과 왜 저럼??????????
print("mean :",scores_lin.mean()) 
print("standard deviation :",scores_lin.std())

scores_forest=cross_val_score(forest_reg,housing_prepared,housing_labels,scoring="neg_mean_squared_error",cv=10)
print("scores :",scores_forest) #???? 결과 왜 저럼??????????
print("mean :",scores_forest.mean()) 
print("standard deviation :",scores_forest.std())
