==> part1/ch2_end_to_end_machine_learning_project/7_fine_tuning.ipynb <==
# 2.7 모델 세부 튜닝 (Fine Tuning)

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np

#load_data
housing_prepared=np.load("datasets/housing/housing_prepared.npy")
housing_labels=pd.read_csv("datasets/housing/housing_labels.csv",usecols=[1],squeeze=True)
print(housing_labels.shape)
print(housing_prepared.shape)

### 1) Grid Search
- GridSearchCv : 탐색하고자 하는 하이퍼파라미터와 시도해볼 값을 지정해주면 가능한 모든 하이퍼파라미터 값에 대해 교차검증을 평가한다

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV

param_grid = [
    # 하이퍼파라미터 12(=3×4)개의 조합을 시도합니다.
    {'n_estimators': [3, 10, 30], 'max_features': [2, 4, 6, 8]},
    # bootstrap은 False로 하고 6(=2×3)개의 조합을 시도합니다.
    {'bootstrap': [False], 'n_estimators': [3, 10], 'max_features': [2, 3, 4]},
  ] # 총 18개 조합 탐색

forest_reg = RandomForestRegressor(random_state=42)
# 다섯 폴드에서 훈련하면 총 (12+6)*5=90번의 훈련이 일어납니다.
grid_search = GridSearchCV(forest_reg, param_grid, cv=5, scoring='neg_mean_squared_error',return_train_score=True, n_jobs=-1)
grid_search.fit(housing_prepared, housing_labels)

grid_search.best_params_

grid_search.best_estimator_

pd.DataFrame(grid_search.cv_results_)

### 2.7.2 Random Search
- Grid search와 달리 가능한 모든 조합을 시도하는 대신 각 반복마다 하이퍼파라미터에 임의의 수를 대입하여 지정한 횟수만큼 평가
- 랜덤 탐색을 1000회 반복실행하면 각 하이퍼파라미터마다 각기 다른 1000개의 값을 탐색

from sklearn.model_selection import RandomizedSearchCV
from scipy.stats import randint

param_distribs = {
        'n_estimators': randint(low=1, high=200),
        'max_features': randint(low=1, high=8),
    }

forest_reg = RandomForestRegressor(random_state=42)
rnd_search = RandomizedSearchCV(forest_reg, param_distributions=param_distribs,
                                n_iter=10, cv=5, scoring='neg_mean_squared_error', 
                                random_state=42, n_jobs=-1)
rnd_search.fit(housing_prepared, housing_labels)

### 2.7.3 Ensemble Learning
- 최상의 모델을 연결해보는 방법

### 2.7.3 Best Model & Error Analysis

# 각 특성의 중요도
feature_importances = grid_search.best_estimator_.feature_importances_
feature_importances

### 2.7.5 System Test

from sklearn.externals import joblib
strat_test_set=pd.read_csv("datasets/housing/strat_test_set.csv",index_col=0)
full_pipeline = joblib.load("full_pipeline.pkl")

final_model = grid_search.best_estimator_

X_test = strat_test_set.drop("median_house_value", axis=1)
y_test = strat_test_set["median_house_value"].copy()

X_test_prepared = full_pipeline.transform(X_test)
final_predictions = final_model.predict(X_test_prepared)

final_mse = mean_squared_error(y_test, final_predictions)
final_rmse = np.sqrt(final_mse)
